Huge refactoring testing
